services/instagram/requirements.txt
instagrapi>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pillow>=10.0.0
python-multipart>=0.0.9
